Boxes File Created Separate boxes.js file and pasted in the boxes section.
