+++ src/App.js
@@ -1,4 +1,5 @@
 import "./styles.css";
+import Box from "./boxes";
 
 function SayHello({ firstName, lastName }) {
   return (
@@ -18,17 +19,6 @@ function CharacterCount({ text }) {
       {" characters"}
     </div>
   );
-}
-
-function Box({ className = "", style, size, ...rest }) {
-  const sizeClassName = size ? `box--${size}` : "box--large";
-
-  const props = {
-    className: `box ${sizeClassName}`,
-    style: { fontStyle: "normal", fontWeight: "600", color: "#fff", ...style },
-    ...rest
-  };
-  return <div {...props} />;
 }
 
 const element = (
